validando todo en problema

+++ aceru-backend/src/main/java/com/aceruservicios/repository/CompetenciaRepo.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.aceruservicios.entity.*;
 import com.aceruservicios.enums.TipoEstadoCompetencia;
 
-public interface CompetenciaRepo extends JpaRepository<Competencia, Integer> {
+public interface CompetenciaRepo extends JpaRepository<Competencia, Long> {
     
     List<Competencia> findByEstadoOrderByFechaFinalDescIdDesc(TipoEstadoCompetencia estado);
 }
